+++ src/components/place-card/current-card.tsx
@@ -1,7 +1,8 @@
-import { PLACES } from '../../const';
+import { PLACES } from '../../mocks/offers';
 import PlaceCard from './place-card';
+import { LOCATIONS } from '../../const';
 
-export default function CurrentPlaces({ city }: { city: string }): JSX.Element {
+export default function CurrentPlaces({ city }: { city: LOCATIONS }): JSX.Element {
   const currentPlaces = PLACES.filter((place) => place.city === city);
 
   return (
