+++ src/components/place-card/favorite-card.tsx
@@ -1,7 +1,8 @@
-import { PLACES } from '../../const';
+import { PLACES } from '../../mocks/offers';
 import PlaceCard from './place-card';
+import { LOCATIONS } from '../../const';
 
-export default function FavoritePlaces({ city }: { city: string }): JSX.Element {
+export default function FavoritePlaces({ city }: { city: LOCATIONS }): JSX.Element {
   const favoritePlaces = PLACES.filter((place) => place.isBookmarked && place.city === city);
 
   return (
